# Filter on journey information efficiently