This brings back the functionality that was reverted in #1800 but with an optimised query making use of the `UNION` operator. This ensures the query planner is able to generate a better query rather than doing a sequential scan on the moves table, it can rely on in indexes.

To actually generate the `UNION` query in Rails, I've followed the code used from [this article](https://joshfrankel.me/blog/a-journey-into-writing-union-queries-with-active-record/) as it's not built in to ActiveRecord. The query is a little odd as it uses an `id IN (...)`, but that's more of a quirk of how it works in ActiveRecord and doesn't introduce any major slowdowns. I've concluded that it's better to keep the query in Ruby rather than dropping down to SQL, than to produce the most optimal query.

I'm confident this could be optimised further, perhaps by doing the `UNION` earlier in the filters, but I think for now this is good enough.

[Jira Ticket](https://dsdmoj.atlassian.net/browse/P4-3339)

## Old query

<details>
  <summary>Click to expand</summary>

```sql
    SELECT
        *     
    FROM
        moves     
    WHERE
        (
            NOT (EXISTS(                               SELECT
                *                               
            FROM
                journeys                               
            WHERE
                state NOT IN ('rejected', 'cancelled')                                 
                AND move_id = moves.id                                 
                AND NOT date = moves.date                           ))                       
            AND date BETWEEN '2022-01-01' AND '2022-01-01'                       
            AND from_location_id = 'aaa4a8b2-c8b1-4bf4-a6c7-8dd67c666bc1'                    
            OR EXISTS(
                SELECT
                    *                               
                FROM
                    journeys                               
                WHERE
                    state NOT IN (
                        'rejected', 'cancelled'
                    )                                 
                    AND move_id = moves.id                                 
                    AND NOT date = moves.date                           
            )                   
            AND EXISTS(
                SELECT
                    *                               
                FROM
                    journeys                               
                WHERE
                    state NOT IN (
                        'rejected', 'cancelled'
                    )                                 
                    AND move_id = moves.id                                 
                    AND date BETWEEN '2022-01-01' AND '2022-01-01'                                 
                    AND from_location_id = 'aaa4a8b2-c8b1-4bf4-a6c7-8dd67c666bc1'                           
            )               
        );
```
</details>

## Old query plan

<details>
  <summary>Click to expand</summary>

```
Seq Scan on moves  (cost=0.00..24512724.04 rows=241374 width=323) (actual time=25227.717..25227.720 rows=0 loops=1)
  Filter: (((NOT (SubPlan 1)) AND (date >= '2022-01-01'::date) AND (date <= '2022-01-01'::date) AND (from_location_id = 'aaa4a8b2-c8b1-4bf4-a6c7-8dd67c666bc1'::uuid)) OR ((SubPlan 2) AND (alternatives: SubPlan 3 or hashed SubPlan 4)))
  Rows Removed by Filter: 956514
  SubPlan 1
    ->  Index Scan using index_journeys_on_move_id_and_client_timestamp on journeys  (cost=0.42..8.45 rows=1 width=0) (actual time=0.022..0.022 rows=0 loops=956514)
          Index Cond: (move_id = moves.id)
          Filter: (((state)::text <> ALL ('{rejected,cancelled}'::text[])) AND (date <> moves.date))
          Rows Removed by Filter: 1
  SubPlan 2
    ->  Index Scan using index_journeys_on_move_id_and_client_timestamp on journeys journeys_1  (cost=0.42..8.45 rows=1 width=0) (actual time=0.002..0.002 rows=0 loops=956514)
          Index Cond: (move_id = moves.id)
          Filter: (((state)::text <> ALL ('{rejected,cancelled}'::text[])) AND (date <> moves.date))
          Rows Removed by Filter: 1
  SubPlan 3
    ->  Index Scan using index_journeys_on_date on journeys journeys_2  (cost=0.42..8.45 rows=1 width=0) (never executed)
          Index Cond: ((date >= '2022-01-01'::date) AND (date <= '2022-01-01'::date))
          Filter: (((state)::text <> ALL ('{rejected,cancelled}'::text[])) AND (move_id = moves.id) AND (from_location_id = 'aaa4a8b2-c8b1-4bf4-a6c7-8dd67c666bc1'::uuid))
  SubPlan 4
    ->  Index Scan using index_journeys_on_date on journeys journeys_3  (cost=0.42..8.45 rows=1 width=16) (actual time=317.856..317.856 rows=0 loops=1)
          Index Cond: ((date >= '2022-01-01'::date) AND (date <= '2022-01-01'::date))
          Filter: (((state)::text <> ALL ('{rejected,cancelled}'::text[])) AND (from_location_id = 'aaa4a8b2-c8b1-4bf4-a6c7-8dd67c666bc1'::uuid))
          Rows Removed by Filter: 475
Planning time: 1.297 ms
Execution time: 25227.794 ms
```
</details>

## New query

<details>
  <summary>Click to expand</summary>

```sql
    SELECT
        "moves".* 
    FROM
        "moves" 
    WHERE
        "moves"."id" IN (
            SELECT
                "moves"."id" 
            FROM
                ((SELECT
                    "moves".* 
                FROM
                    "moves" 
                WHERE
                    NOT (EXISTS (SELECT
                        "journeys".* 
                    FROM
                        "journeys" 
                    WHERE
                        "journeys"."state" NOT IN ('rejected', 'cancelled') 
                        AND (move_id = moves.id) 
                        AND NOT (date = moves.date))) 
                    AND (
                        date BETWEEN '2022-01-01' AND '2022-01-01'
                    ) 
                    AND "moves"."from_location_id" = '58acecc0-661c-4141-9c42-882af6843e1b'
                ) 
            UNION
            (
                SELECT
                    "moves".* 
                FROM
                    "moves" 
                WHERE
                    EXISTS (
                        SELECT
                            "journeys".* 
                        FROM
                            "journeys" 
                        WHERE
                            "journeys"."state" NOT IN (
                                'rejected', 'cancelled'
                            ) 
                            AND (
                                move_id = moves.id
                            ) 
                            AND NOT (date = moves.date)
                    ) 
                    AND EXISTS (
                        SELECT
                            "journeys".* 
                        FROM
                            "journeys" 
                        WHERE
                            "journeys"."state" NOT IN (
                                'rejected', 'cancelled'
                            ) 
                            AND (
                                move_id = moves.id
                            ) 
                            AND (
                                date BETWEEN '2022-01-01' AND '2022-01-01'
                            ) 
                            AND "journeys"."from_location_id" = '58acecc0-661c-4141-9c42-882af6843e1b'
                    )
                )
            ) moves
    ) 
```
</details>

## New query plan

<details>
  <summary>Click to expand</summary>

```
Sort  (cost=35.17..35.23 rows=24 width=323) (actual time=0.975..0.980 rows=0 loops=1)
  Sort Key: moves.date DESC
  Sort Method: quicksort  Memory: 25kB
  ->  Append  (cost=0.85..34.62 rows=24 width=323) (actual time=0.960..0.963 rows=0 loops=1)
        ->  Nested Loop Anti Join  (cost=0.85..16.91 rows=1 width=323) (actual time=0.398..0.399 rows=0 loops=1)
              ->  Index Scan using index_moves_on_date on moves  (cost=0.42..8.45 rows=1 width=323) (actual time=0.396..0.397 rows=0 loops=1)
                    Index Cond: ((date >= '2022-01-01'::date) AND (date <= '2022-01-01'::date))
                    Filter: (from_location_id = 'aaa4a8b2-c8b1-4bf4-a6c7-8dd67c666bc1'::uuid)
                    Rows Removed by Filter: 447
              ->  Index Scan using index_journeys_on_move_id_and_client_timestamp on journeys  (cost=0.42..8.45 rows=1 width=20) (never executed)
                    Index Cond: (move_id = moves.id)
                    Filter: (((state)::text <> ALL ('{rejected,cancelled}'::text[])) AND (date <> moves.date))
        ->  Nested Loop Semi Join  (cost=9.30..17.47 rows=23 width=323) (actual time=0.561..0.563 rows=0 loops=1)
              ->  Nested Loop  (cost=8.88..16.91 rows=1 width=339) (actual time=0.560..0.562 rows=0 loops=1)
                    ->  HashAggregate  (cost=8.45..8.46 rows=1 width=16) (actual time=0.560..0.560 rows=0 loops=1)
                          Group Key: journeys_2.move_id
                          ->  Index Scan using index_journeys_on_date on journeys journeys_2  (cost=0.42..8.45 rows=1 width=16) (actual time=0.558..0.558 rows=0 loops=1)
                                Index Cond: ((date >= '2022-01-01'::date) AND (date <= '2022-01-01'::date))
                                Filter: (((state)::text <> ALL ('{rejected,cancelled}'::text[])) AND (from_location_id = 'aaa4a8b2-c8b1-4bf4-a6c7-8dd67c666bc1'::uuid))
                                Rows Removed by Filter: 475
                    ->  Index Scan using moves_pkey on moves moves_1  (cost=0.42..8.44 rows=1 width=323) (never executed)
                          Index Cond: (id = journeys_2.move_id)
              ->  Index Scan using index_journeys_on_move_id on journeys journeys_1  (cost=0.42..0.56 rows=1 width=20) (never executed)
                    Index Cond: (move_id = moves_1.id)
                    Filter: (((state)::text <> ALL ('{rejected,cancelled}'::text[])) AND (date <> moves_1.date))
Planning time: 1.378 ms
Execution time: 1.107 ms
```
</details>